+++ ecs/ecs.h
@@ -7,11 +7,11 @@
 #include "dtypes.h"
 #include "log.h"
 
-#include "collider.h"
-#include "texture.h"
-#include "actor.h"
-#include "tilemap.h"
-#include "light.h"
+#include "c_collider.h"
+#include "c_texture.h"
+#include "c_actor.h"
+#include "c_tilemap.h"
+#include "c_light.h"
 
 #define INVALID_ENTITY Entity::createID(EntityIndex(-1), 0)