+++ ecs/systems/s_actor.h
@@ -7,6 +7,7 @@
 #include "dtypes.h"
 #include "log.h"
 
+#include "ecs.h"
 #include "scene.h"
 
 namespace Verse::System::Actor