+++ graphics/r_shader.h
@@ -15,28 +15,6 @@
 #include <fstream>
 #include <streambuf>
 
-#include <SDL2/SDL.h>
-#include <SDL2/SDL_image.h>
-
-#ifdef __APPLE__ //MACOS
-
-    #define GL_SILENCE_DEPRECATION
-
-    #include <OpenGL/OpenGL.h>
-
-    #if ESSENTIAL_GL_PRACTICES_SUPPORT_GL3
-        #include <OpenGL/gl3.h>
-    #else
-        #include <OpenGL/gl.h>
-    #endif //!ESSENTIAL_GL_PRACTICES_SUPPORT_GL3
-
-    #else //OTHERS (Not tested)
-
-    #include <SDL2/SDL_opengl.h>
-    #include <SDL2/SDL_opengl_glext.h>
-
-#endif //!__APPLE__
-
 namespace Verse::Graphics::Shader
 {
     bool initGLExtensions();